Use effient float (not double) trig functions

+++ src/MadgwickAHRS.h
@@ -37,9 +37,9 @@ public:
     Madgwick(void);
     void update(float gx, float gy, float gz, float ax, float ay, float az, float mx, float my, float mz);
     void updateIMU(float gx, float gy, float gz, float ax, float ay, float az);
-    float getPitch(){return atan2(2 * q2 * q3 - 2 * q0 * q1, 2 * q0 * q0 + 2 * q3 * q3 - 1);};
-    float getRoll(){return -1 * asin(2 * q1 * q3 + 2 * q0 * q2);};
-    float getYaw(){return atan2(2 * q1 * q2 - 2 * q0 * q3, 2 * q0 * q0 + 2 * q1 * q1 - 1);};
+    float getPitch(){return atan2f(2 * q2 * q3 - 2 * q0 * q1, 2 * q0 * q0 + 2 * q3 * q3 - 1);};
+    float getRoll(){return -1 * asinf(2 * q1 * q3 + 2 * q0 * q2);};
+    float getYaw(){return atan2f(2 * q1 * q2 - 2 * q0 * q3, 2 * q0 * q0 + 2 * q1 * q1 - 1);};
 };
 #endif
